# docs/swagger.yaml
basePath: /
definitions:
...
  groups.AssignRolesRequest:
    properties:
      role_ids:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
    type: object
...
  groups.ErrorResponse:
    properties:
      error:
        example: Bad Request
        type: string
      message:
        example: Invalid group name
        type: string
    type: object
...
  groups.GroupRoleResponse:
    properties:
      group_public_id:
        example: grp-abc123
        type: string
      role_id:
        example: 1
        type: integer
      role_name:
        example: admin
        type: string
    type: object
...
  groups.SuccessResponse:
    properties:
      message:
        example: Group deleted successfully
        type: string
    type: object
...
  roles.AssignUserRolesRequest:
    properties:
      role_ids:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
    type: object
  roles.BatchCreateRoleRequest:
    properties:
      roles:
        items:
          $ref: '#/definitions/roles.CreateRoleRequest'
        type: array
    type: object
  roles.BatchDeleteRoleRequest:
    properties:
      ids:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
    type: object
  roles.BatchOperationResponse:
    properties:
      failed_count:
        example: 0
        type: integer
      failed_ids:
        items:
          type: integer
        type: array
      success_count:
        example: 5
        type: integer
    type: object
  roles.BatchUpdateRoleRequest:
    properties:
      updates:
        items:
          properties:
            description:
              type: object
            id:
              example: 1
              type: integer
            name:
              example: admin
              type: string
          type: object
        type: array
    type: object
  roles.CreateRoleRequest:
    properties:
      description:
        type: object
      name:
        example: admin
        type: string
    type: object
  roles.ErrorResponse:
    properties:
      error:
        example: Bad Request
        type: string
      message:
        example: Invalid role name
        type: string
    type: object
  roles.RoleListResponseWrapper:
    properties:
      data:
        items:
          $ref: '#/definitions/roles.RoleResponse'
        type: array
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total_count:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  roles.RoleResponse:
    properties:
      description:
        type: object
      id:
        example: 1
        type: integer
      name:
        example: admin
        type: string
    type: object
  roles.SearchRoleRequest:
    properties:
      description:
        example: administrator
        type: string
      name:
        example: admin
        type: string
    type: object
  roles.SuccessResponse:
    properties:
      message:
        example: Role deleted successfully
        type: string
    type: object
  roles.UpdateRoleRequest:
    properties:
      description:
        type: object
      name:
        example: admin
        type: string
    type: object
  roles.UserRoleResponse:
    properties:
      role_id:
        example: 1
        type: integer
      role_name:
        example: admin
        type: string
      user_id:
        example: 1
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@knowledgecenter.io
    name: API Support
  description: Knowledge Center REST API Server
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  title: Knowledge Center API
  version: "1.0"
paths:
...
  /groups/{publicId}/roles:
    get:
      consumes:
      - application/json
      description: Retrieves all roles assigned to a group
      parameters:
      - description: Group Public ID
        in: path
        name: publicId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/groups.GroupRoleResponse'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get group roles
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Assigns roles to a group (replaces existing roles)
      parameters:
      - description: Group Public ID
        in: path
        name: publicId
        required: true
        type: string
      - description: Role IDs to assign
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/groups.AssignRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/groups.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assign roles to group
      tags:
      - groups
  /groups/{publicId}/roles/{roleId}:
    delete:
      consumes:
      - application/json
      description: Removes a role from a group
      parameters:
      - description: Group Public ID
        in: path
        name: publicId
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/groups.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/groups.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove role from group
      tags:
      - groups
...
  /roles:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of roles
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roles.RoleListResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Creates a new role with unique name
      parameters:
      - description: Role data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/roles.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/roles.RoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
        "409":
          description: Role name already exists
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new role
      tags:
      - roles
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roles.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete role
      tags:
      - roles
    get:
      consumes:
      - application/json
      description: Retrieves a role by its ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roles.RoleResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get role by ID
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Updates an existing role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role data to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/roles.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roles.RoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
        "409":
          description: Role name already exists
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - roles
  /roles/{id}/users:
    get:
      consumes:
      - application/json
      description: Retrieves all user IDs that have a specific role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get users with role
      tags:
      - roles
  /roles/batch:
    delete:
      consumes:
      - application/json
      description: Deletes multiple roles in a single transaction
      parameters:
      - description: Batch delete request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/roles.BatchDeleteRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roles.BatchOperationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch delete roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Creates multiple roles in a single transaction
      parameters:
      - description: Batch create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/roles.BatchCreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/roles.BatchOperationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch create roles
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Updates multiple roles in a single transaction
      parameters:
      - description: Batch update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/roles.BatchUpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roles.BatchOperationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch update roles
      tags:
      - roles
  /roles/search:
    post:
      consumes:
      - application/json
      description: Searches for roles based on name or description
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search criteria
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/roles.SearchRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roles.RoleListResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search roles
      tags:
      - roles
...
  /users/{userId}/roles:
    get:
      consumes:
      - application/json
      description: Retrieves all roles assigned to a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/roles.UserRoleResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user roles
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Assigns roles to a user (replaces existing roles)
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Role IDs to assign
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/roles.AssignUserRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roles.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assign roles to user
      tags:
      - roles
  /users/{userId}/roles/{roleId}:
    delete:
      consumes:
      - application/json
      description: Removes a specific role from a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roles.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/roles.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove role from user
      tags:
      - roles
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"